Move to collapsibleSidebar to UI redux key
