+++ src/ui/actions.js
@@ -1,0 +1,14 @@
+// Action labels
+export const actionsList = {
+  collapsibleSidebar: "TOGGLE_COLLAPSIBLE_SIDEBAR"
+};
+
+/**
+  Toggle collapsible sidebar
+*/
+export function collapsibleSidebar(toggle) {
+  return {
+    type: actionsList.collapsibleSidebar,
+    payload: toggle
+  };
+}
