+++ src/ui/selectors.js
@@ -1,0 +1,15 @@
+// ----- Exported functions
+
+/**
+  Checks state to see if the sidebar is collapsible.
+  Usage:
+    const isSidebarCollapsible = useSelector(isSidebarCollapsible);
+
+  @param {Object} state The application state.
+  @returns {Boolean} If the sidebar is collapsible.
+*/
+export const isSidebarCollapsible = state => {
+  if (state && state.ui) {
+    return state.ui.collapsibleSidebar;
+  }
+};
